Add a lowercaser for binary creation check

+++ assembler-linux/lib/compile.sh
@@ -1,9 +1,12 @@
-#
+#!/bin/sh
 # This script assumes that:
 # $FN is set to the assembler source filename (without extension)
 # $TMP_CONF is a synthesized dosbox.conf file
 # $DOSBOX_OUTPUT tells DOSBox where to dump stdout 
 # $DOSBOX_EXIT tells DOSBox if it needs to automatically close itself
+
+[ -z $FN ] && FN=$1
+[ -z $FN ] && exit 41
 
 # Cleaning binaries
 [ -f $FN.exe ] && rm $FN.exe
